Fix: 質問作成時の画像プレビュー機能の修正
