+++ app/views/questions/_form.html.erb
@@ -2,7 +2,7 @@
   <%= render 'shared/error_message', object: @question %>
   <div class="form-group">
     <%= f.label :photo %></br>
-    <%= f.file_field :photo, onchange: 'previewImage()', class: 'assign-form mb-3', accept: 'image/*' %>
+    <%= f.file_field :photo, class: 'assign-form mb-3', accept: 'image/*', id: 'atction-button' %>
     <%= f.hidden_field :photo_cache %>
     <div class='mt-3 mb-3'>
       <%= image_tag question.photo.url,
